29 refactor shell grammar to mimic bash

+++ lib/mem/ft_memset.c
@@ -8,8 +8,7 @@ void *ft_memset(void *str, const int c, size_t n)
 	p = (byte *)str;
 	while (n-- > 0)
 	{
-		// printf("Copying... %zu\n", n);
-		*p++ = c;
+		*p++ = (byte)c;
 	}
 	return str;
 }
